Remove use of Context for 'readonly'
We need to make the readonly-ness of the table available to the
AcqInvoiceLineitemView component. It is supplied by the calling
function. We initially used the Context API for this. But as detailed in
AcqInvoiceLineItemActions.test.js, this currently affects its
testability. Since AcqInvoiceLineitemView is only one level deep in the
component tree, we now pass the 'readonly' property via props instead